0.0.76 Pre-alpha - Added new rights: Edit your own comments, Create new comments, Superuser; - Fixed a bug where a comment was created under a user from a previous session if it was launched under the same IP address; - Fixed a bug where the comment management panel on the full post view page on the website was not assembled correctly; - Fixed user initialization during authentication. Previously, the system returned user data from the previous session launched from the same IP address; - Fixed a bug where it was impossible to change the password, because when checking the old password, the system in any case considered that the old password was incorrect; - Fixed a bug where it was possible to change the rating of self own comment; - Fixed stylistic errors in the default website template; - Localizations updated; - Minor bug fixes;

+++ api/entry/comment/delete.handler.php
@@ -28,7 +28,7 @@ if ($system_core->client->is_logged(1) || $system_core->client->is_logged(2)) {
       $comment = new EntryComment($system_core, $comment_id);
       $comment->init_data(['author_id']);
 
-      if ($client_user_group->permission_check($client_user_group::PERMISSION_MODER_ENTRIES_COMMENTS_MANAGEMENT) || $comment->get_author_id() == $client_user->get_id()) {
+      if ($client_user_group->permission_check($client_user_group::PERMISSION_MODER_ENTRIES_COMMENTS_MANAGEMENT) || $client_user_group->permission_check($client_user_group::PERMISSION_BASE_ENTRY_COMMENT_CHANGE) && $comment->get_author_id() == $client_user->get_id()) {
         $comment_is_deleted = $comment->delete();
 
         if ($comment_is_deleted) {
